Fix bug with task id being a int instead of a long

+++ src/main/kotlin/com/davidcorrado/kotlinrest/data/Task.kt
@@ -10,7 +10,7 @@ class Task() {
 
     @Id
     @GeneratedValue(strategy = GenerationType.AUTO)
-    val id: Int? = 0
+    val id: Long? = 0
     var name: String = ""
     var detail: String = ""
